requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.35.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0
requests>=2.28.0
tqdm>=4.64.0
